Kernel/SysFS: Split bulky SysFSPCI file into separate files
